02  create a second model and move stuff to common
