+++ tests/models/py3_sklearn_2/custom.py
@@ -7,6 +7,8 @@
 A super simple example of custom model body with a single 'transform' hook.
 """
 
+from common import conversion
+
 
 def transform(data, model):
     """
@@ -31,4 +33,9 @@ def transform(data, model):
         if target_col in data:
             data.pop(target_col)
     data = data.fillna(0)
+
+    inches = 10
+    centimeters = conversion.inch_to_cm(inches)
+    print(f"Inches: {inches}, Centimeters: {centimeters}", flush=True)
+
     return data
