+++ tests/models/py3_sklearn_2/model-2.yaml
@@ -1,9 +1,9 @@
-user_provided_model_id: my-awesome-model-id-1-630c72d2562155a4d953492e
+user_provided_model_id: my-awesome-model-id-2-630c72e3562155a879bc1445
 
 target_type: Regression
 
 settings:
-  name: My Awsome GitHub Model 1 [GitHub CI/CD]
+  name: My Awsome GitHub Model 2 [GitHub CI/CD]
   target_name: Grade 2014
 
 version:
@@ -12,6 +12,4 @@ version:
   model_environment_id: 5e8c889607389fe0f466c72d
 
   include_glob_pattern: ["./*", "../common/*"]
-  exclude_glob_pattern: ["README.md", "model*.yaml"]
-
-  memory: 512Mi
+  exclude_glob_pattern: ["model*.yaml"]
